Moved MessageTemplate/Attachment and NetworkerReplyTrigger from static.py to dynamic.py
