+++ mln/models/static.py
@@ -268,32 +268,6 @@ class NetworkerFriendshipConditionSource(models.Model):
 	def __str__(self):
 		return self.source
 
-class MessageTemplate(models.Model): 
-	"""An action that triggers a networker to send a response to the user."""
-	networker = models.ForeignKey(User, null=True, related_name="+", on_delete=models.CASCADE, limit_choices_to={"profile__is_networker": True})
-	response = models.ForeignKey(on_delete=models.deletion.CASCADE, related_name='+', to='mln.messagebody')
-
-	# Currently meant for devs to collect data on triggers, later to be properly integrated into the system.
-	networker_name = models.CharField(max_length=64, blank=True, null=True)
-	trigger = models.TextField(blank=True, null=True)
-	source = models.TextField(blank=True, null=True)
-	notes = models.TextField(blank=True, null=True)
-
-class NetworkerReplyTrigger(MessageTemplate):
-	"""A template for a message that triggers a networker's response."""
-	message_body = models.ForeignKey(MessageBody, related_name="+", on_delete=models.CASCADE, blank=True, null=True)
-	message_attachment = models.ForeignKey(ItemInfo, related_name="+", on_delete=models.CASCADE, blank=True, null=True)
-
-	def __str__(self):
-		return "From %s: %s" % (self.networker, self.response)
-
-class MessageTemplateAttachment(Stack):
-	"""A stack to be attached to a networker message."""
-	template = models.ForeignKey(MessageTemplate, related_name="attachments", on_delete=models.CASCADE)
-
-	class Meta:
-		constraints = (models.UniqueConstraint(fields=("template", "item"), name="networker_template_attachment_unique_template_item"),)
-
 class NetworkerPageSource(models.Model):
 	"""Documentation only: Note whether a reconstructed networker page has a graphical source, or is tentatively reconstructed from a description."""
 	networker = models.OneToOneField(User, related_name="+", on_delete=models.CASCADE, limit_choices_to={"profile__is_networker": True})
